Preparing package simplification
