+++ src/Bundle/JoseFramework/DependencyInjection/Source/Encryption/EncryptionSource.php
@@ -10,10 +10,8 @@ use Jose\Bundle\JoseFramework\DependencyInjection\Source\Source;
 use Jose\Bundle\JoseFramework\DependencyInjection\Source\SourceWithCompilerPasses;
 use Jose\Component\Encryption\Algorithm\ContentEncryption\AESCBCHS;
 use Jose\Component\Encryption\Algorithm\ContentEncryption\AESGCM;
-use Jose\Component\Encryption\Algorithm\KeyEncryption\A128CTR;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\AESGCMKW;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\AESKW;
-use Jose\Component\Encryption\Algorithm\KeyEncryption\Chacha20Poly1305;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\Dir;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\ECDHES;
 use Jose\Component\Encryption\Algorithm\KeyEncryption\PBES2AESKW;
@@ -21,6 +19,8 @@ use Jose\Component\Encryption\Algorithm\KeyEncryption\RSA;
 use Jose\Component\Encryption\JWEBuilderFactory;
 use Jose\Component\Encryption\JWEDecrypterFactory;
 use Jose\Component\Encryption\Serializer\JWESerializer as JWESerializerAlias;
+use Jose\Experimental\KeyEncryption\A128CTR;
+use Jose\Experimental\KeyEncryption\Chacha20Poly1305;
 use Symfony\Component\Config\Definition\Builder\NodeDefinition;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
