+++ src/Bundle/JoseFramework/Resources/config/Algorithms/encryption_experimental_chacha20_poly1305.php
@@ -2,7 +2,8 @@
 
 declare(strict_types=1);
 
-use Jose\Component\Encryption\Algorithm\KeyEncryption\Chacha20Poly1305;
+use Jose\Experimental\KeyEncryption\Chacha20Poly1305;
+use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 /*
  * The MIT License (MIT)
@@ -12,8 +13,6 @@ use Jose\Component\Encryption\Algorithm\KeyEncryption\Chacha20Poly1305;
  * This software may be modified and distributed under the terms
  * of the MIT license.  See the LICENSE file for details.
  */
-
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 /*
  * ---- New algorithms ----
