+++ src/Bundle/JoseFramework/Resources/config/Algorithms/signature_experimental.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-use Jose\Component\Signature\Algorithm\Blake2b;
-use Jose\Component\Signature\Algorithm\ES256K;
-use Jose\Component\Signature\Algorithm\HS1;
-use Jose\Component\Signature\Algorithm\HS256_64;
-use Jose\Component\Signature\Algorithm\RS1;
+use Jose\Experimental\Signature\Blake2b;
+use Jose\Experimental\Signature\ES256K;
+use Jose\Experimental\Signature\HS1;
+use Jose\Experimental\Signature\HS256_64;
+use Jose\Experimental\Signature\RS1;
+use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 /*
  * The MIT License (MIT)
@@ -16,8 +17,6 @@ use Jose\Component\Signature\Algorithm\RS1;
  * This software may be modified and distributed under the terms
  * of the MIT license.  See the LICENSE file for details.
  */
-
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 /*
  * ---- New algorithms ----
