+++ src/EncryptionAlgorithm/Experimental/composer.json
@@ -32,14 +32,8 @@
             "homepage": "https://github.com/web-token/jwt-framework/contributors"
         }
     ],
-    "autoload": {
-        "psr-4": {
-            "Jose\\Component\\Encryption\\Algorithm\\": ""
-        }
-    },
+    "autoload": {},
     "require": {
-        "php": ">=8.1",
-        "ext-openssl": "*",
-        "web-token/jwt-encryption-algorithm-rsa": "^3.2"
+        "web-token/jwt-experimental": "^3.2"
     }
 }
