+++ src/EncryptionAlgorithm/KeyEncryption/AESKW/composer.json
@@ -33,5 +33,7 @@
         }
     ],
     "autoload": {},
-    "require": {}
+    "require": {
+        "web-token/jwt-encryption": "^3.2"
+    }
 }
