+++ src/Experimental/ContentEncryption/A128CCM_16_64.php
@@ -2,18 +2,18 @@
 
 declare(strict_types=1);
 
-namespace Jose\Component\Encryption\Algorithm\ContentEncryption;
+namespace Jose\Experimental\ContentEncryption;
 
-final class A256CCM_16_128 extends AESCCM
+final class A128CCM_16_64 extends AESCCM
 {
     public function getCEKSize(): int
     {
-        return 256;
+        return 128;
     }
 
     public function name(): string
     {
-        return 'A256CCM-16-128';
+        return 'A128CCM-16-64';
     }
 
     public function getIVSize(): int
@@ -23,7 +23,7 @@ final class A256CCM_16_128 extends AESCCM
 
     protected function getMode(): string
     {
-        return 'aes-256-ccm';
+        return 'aes-128-ccm';
     }
 
     protected function getTagLength(): int
