+++ src/Experimental/ContentEncryption/A128CCM_64_128.php
@@ -2,32 +2,32 @@
 
 declare(strict_types=1);
 
-namespace Jose\Component\Encryption\Algorithm\ContentEncryption;
+namespace Jose\Experimental\ContentEncryption;
 
-final class A256CCM_16_64 extends AESCCM
+final class A128CCM_64_128 extends AESCCM
 {
     public function getCEKSize(): int
     {
-        return 256;
+        return 128;
     }
 
     public function name(): string
     {
-        return 'A256CCM-16-64';
+        return 'A128CCM-64-128';
     }
 
     public function getIVSize(): int
     {
-        return 13 * 8;
+        return 7 * 8;
     }
 
     protected function getMode(): string
     {
-        return 'aes-256-ccm';
+        return 'aes-128-ccm';
     }
 
     protected function getTagLength(): int
     {
-        return 8;
+        return 16;
     }
 }
