+++ src/Experimental/ContentEncryption/A128CCM_64_64.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Jose\Component\Encryption\Algorithm\ContentEncryption;
+namespace Jose\Experimental\ContentEncryption;
 
 final class A128CCM_64_64 extends AESCCM
 {
