+++ src/Experimental/KeyEncryption/AESCTR.php
@@ -2,34 +2,22 @@
 
 declare(strict_types=1);
 
-namespace Jose\Component\Encryption\Algorithm\KeyEncryption;
+namespace Jose\Experimental\KeyEncryption;
 
 use InvalidArgumentException;
 use Jose\Component\Core\JWK;
-use LogicException;
+use Jose\Component\Encryption\Algorithm\KeyEncryption\KeyEncryption;
 use ParagonIE\ConstantTime\Base64UrlSafe;
 use RuntimeException;
 use function in_array;
 use function is_string;
 use const OPENSSL_RAW_DATA;
 
-final class Chacha20Poly1305 implements KeyEncryption
+abstract class AESCTR implements KeyEncryption
 {
-    public function __construct()
-    {
-        if (! in_array('chacha20-poly1305', openssl_get_cipher_methods(), true)) {
-            throw new LogicException('The algorithm "chacha20-poly1305" is not supported in this platform.');
-        }
-    }
-
     public function allowedKeyTypes(): array
     {
         return ['oct'];
-    }
-
-    public function name(): string
-    {
-        return 'chacha20-poly1305';
     }
 
     /**
@@ -39,14 +27,13 @@ final class Chacha20Poly1305 implements KeyEncryption
     public function encryptKey(JWK $key, string $cek, array $completeHeader, array &$additionalHeader): string
     {
         $k = $this->getKey($key);
-        $nonce = random_bytes(12);
+        $iv = random_bytes(16);
 
         // We set header parameters
-        $additionalHeader['nonce'] = Base64UrlSafe::encodeUnpadded($nonce);
+        $additionalHeader['iv'] = Base64UrlSafe::encodeUnpadded($iv);
 
-        $tag = null;
-        $result = openssl_encrypt($cek, 'chacha20-poly1305', $k, OPENSSL_RAW_DATA, $nonce, $tag);
-        if ($result === false || ! is_string($tag)) {
+        $result = openssl_encrypt($cek, $this->getMode(), $k, OPENSSL_RAW_DATA, $iv);
+        if ($result === false) {
             throw new RuntimeException('Unable to encrypt the CEK');
         }
 
@@ -59,14 +46,11 @@ final class Chacha20Poly1305 implements KeyEncryption
     public function decryptKey(JWK $key, string $encrypted_cek, array $header): string
     {
         $k = $this->getKey($key);
-        isset($header['nonce']) || throw new InvalidArgumentException('The header parameter "nonce" is missing.');
-        is_string($header['nonce']) || throw new InvalidArgumentException('The header parameter "nonce" is not valid.');
-        $nonce = Base64UrlSafe::decode($header['nonce']);
-        if (mb_strlen($nonce, '8bit') !== 12) {
-            throw new InvalidArgumentException('The header parameter "nonce" is not valid.');
-        }
+        isset($header['iv']) || throw new InvalidArgumentException('The header parameter "iv" is missing.');
+        is_string($header['iv']) || throw new InvalidArgumentException('The header parameter "iv" is not valid.');
+        $iv = Base64UrlSafe::decode($header['iv']);
 
-        $result = openssl_decrypt($encrypted_cek, 'chacha20-poly1305', $k, OPENSSL_RAW_DATA, $nonce);
+        $result = openssl_decrypt($encrypted_cek, $this->getMode(), $k, OPENSSL_RAW_DATA, $iv);
         if ($result === false) {
             throw new RuntimeException('Unable to decrypt the CEK');
         }
@@ -78,6 +62,8 @@ final class Chacha20Poly1305 implements KeyEncryption
     {
         return self::MODE_ENCRYPT;
     }
+
+    abstract protected function getMode(): string;
 
     private function getKey(JWK $key): string
     {
