+++ src/Experimental/KeyEncryption/Chacha20Poly1305.php
@@ -2,21 +2,35 @@
 
 declare(strict_types=1);
 
-namespace Jose\Component\Encryption\Algorithm\KeyEncryption;
+namespace Jose\Experimental\KeyEncryption;
 
 use InvalidArgumentException;
 use Jose\Component\Core\JWK;
+use Jose\Component\Encryption\Algorithm\KeyEncryption\KeyEncryption;
+use LogicException;
 use ParagonIE\ConstantTime\Base64UrlSafe;
 use RuntimeException;
 use function in_array;
 use function is_string;
 use const OPENSSL_RAW_DATA;
 
-abstract class AESCTR implements KeyEncryption
+final class Chacha20Poly1305 implements KeyEncryption
 {
+    public function __construct()
+    {
+        if (! in_array('chacha20-poly1305', openssl_get_cipher_methods(), true)) {
+            throw new LogicException('The algorithm "chacha20-poly1305" is not supported in this platform.');
+        }
+    }
+
     public function allowedKeyTypes(): array
     {
         return ['oct'];
+    }
+
+    public function name(): string
+    {
+        return 'chacha20-poly1305';
     }
 
     /**
@@ -26,13 +40,14 @@ abstract class AESCTR implements KeyEncryption
     public function encryptKey(JWK $key, string $cek, array $completeHeader, array &$additionalHeader): string
     {
         $k = $this->getKey($key);
-        $iv = random_bytes(16);
+        $nonce = random_bytes(12);
 
         // We set header parameters
-        $additionalHeader['iv'] = Base64UrlSafe::encodeUnpadded($iv);
+        $additionalHeader['nonce'] = Base64UrlSafe::encodeUnpadded($nonce);
 
-        $result = openssl_encrypt($cek, $this->getMode(), $k, OPENSSL_RAW_DATA, $iv);
-        if ($result === false) {
+        $tag = null;
+        $result = openssl_encrypt($cek, 'chacha20-poly1305', $k, OPENSSL_RAW_DATA, $nonce, $tag);
+        if ($result === false || ! is_string($tag)) {
             throw new RuntimeException('Unable to encrypt the CEK');
         }
 
@@ -45,11 +60,14 @@ abstract class AESCTR implements KeyEncryption
     public function decryptKey(JWK $key, string $encrypted_cek, array $header): string
     {
         $k = $this->getKey($key);
-        isset($header['iv']) || throw new InvalidArgumentException('The header parameter "iv" is missing.');
-        is_string($header['iv']) || throw new InvalidArgumentException('The header parameter "iv" is not valid.');
-        $iv = Base64UrlSafe::decode($header['iv']);
+        isset($header['nonce']) || throw new InvalidArgumentException('The header parameter "nonce" is missing.');
+        is_string($header['nonce']) || throw new InvalidArgumentException('The header parameter "nonce" is not valid.');
+        $nonce = Base64UrlSafe::decode($header['nonce']);
+        if (mb_strlen($nonce, '8bit') !== 12) {
+            throw new InvalidArgumentException('The header parameter "nonce" is not valid.');
+        }
 
-        $result = openssl_decrypt($encrypted_cek, $this->getMode(), $k, OPENSSL_RAW_DATA, $iv);
+        $result = openssl_decrypt($encrypted_cek, 'chacha20-poly1305', $k, OPENSSL_RAW_DATA, $nonce);
         if ($result === false) {
             throw new RuntimeException('Unable to decrypt the CEK');
         }
@@ -61,8 +79,6 @@ abstract class AESCTR implements KeyEncryption
     {
         return self::MODE_ENCRYPT;
     }
-
-    abstract protected function getMode(): string;
 
     private function getKey(JWK $key): string
     {
