+++ src/Experimental/Signature/RS1.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Jose\Component\Signature\Algorithm;
+namespace Jose\Experimental\Signature;
+
+use Jose\Component\Signature\Algorithm\RSAPKCS1;
 
 final class RS1 extends RSAPKCS1
 {
