+++ src/SignatureAlgorithm/Experimental/composer.json
@@ -32,14 +32,8 @@
             "homepage": "https://github.com/web-token/jwt-framework/contributors"
         }
     ],
-    "autoload": {
-        "psr-4": {
-            "Jose\\Component\\Signature\\Algorithm\\": ""
-        }
-    },
+    "autoload": {},
     "require": {
-        "php": ">=8.1",
-        "web-token/jwt-signature-algorithm-rsa": "^3.2",
-        "web-token/jwt-signature-algorithm-hmac": "^3.2"
+        "web-token/jwt-experimental": "^3.2"
     }
 }
