+++ tests/EncryptionAlgorithm/Experimental/Chacha20Poly1305ContentEncryptionTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Jose\Tests\Component\Encryption\Algorithm;
 
 use Jose\Component\Core\JWK;
-use Jose\Component\Encryption\Algorithm\KeyEncryption\Chacha20Poly1305;
+use Jose\Experimental\KeyEncryption\Chacha20Poly1305;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Throwable;
