+++ tests/SignatureAlgorithm/Experimental/P256KSignatureTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Jose\Tests\Component\Signature\Algorithm;
 
 use Jose\Component\Core\JWK;
-use Jose\Component\Signature\Algorithm\ES256K;
+use Jose\Experimental\Signature\ES256K;
 use ParagonIE\ConstantTime\Base64UrlSafe;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
